# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carteira Financeira API",
        "description": "API para gerenciamento de carteira financeira",
        "contact": {
            "name": "Suporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contas/{publicId}": {
            "get": {
                "tags": [
                    "Contas"
                ],
                "summary": "Busca detalhes de uma conta",
                "description": "Retorna os detalhes completos de uma conta específica",
                "operationId": "41fc430090f7e4e885db777a5ce3aa09",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "description": "ID público da conta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da conta retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID público da conta",
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "description": "ID público do usuário",
                                            "type": "string"
                                        },
                                        "balance": {
                                            "description": "Saldo atual da conta",
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "status": {
                                            "description": "Status atual da conta",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Data de criação",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Data de última atualização",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contas/{publicId}/depositar": {
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Realiza um depósito direto na conta",
                "description": "Efetua um depósito direto em uma conta específica",
                "operationId": "e56245e42f35de605642ab40809b5340",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "description": "ID público da conta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Valor do depósito",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Depósito realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Conta"
                                        },
                                        "meta": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/MetaPaginacao"
                                                }
                                            ],
                                            "nullable": true,
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação dos dados"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contas/{publicId}/sacar": {
            "post": {
                "tags": [
                    "Contas"
                ],
                "summary": "Realiza um saque da conta",
                "description": "Efetua um saque de uma conta específica",
                "operationId": "d0bd911a416e291e973cdc557764104b",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "description": "ID público da conta",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Valor do saque",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saque realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Conta"
                                        },
                                        "meta": {
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/MetaPaginacao"
                                                }
                                            ],
                                            "nullable": true,
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou saldo insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Conta": {
                "title": "Conta",
                "description": "Modelo de conta bancária conforme resposta da API",
                "properties": {
                    "id": {
                        "description": "ID público da conta",
                        "type": "string",
                        "example": "c1a2b3c4-d5e6-7890-abcd-1234567890ef"
                    },
                    "user_id": {
                        "description": "ID público do usuário",
                        "type": "string",
                        "example": "u1b2c3d4-e5f6-7890-abcd-1234567890ef"
                    },
                    "balance": {
                        "description": "Saldo atual da conta",
                        "type": "number",
                        "format": "float",
                        "example": 1500.75
                    },
                    "status": {
                        "description": "Status atual da conta. Valores possíveis: 'active', 'inactive'",
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-10T15:30:00Z"
                    }
                },
                "type": "object"
            },
...
            "MetaPaginacao": {
                "title": "Metadados de Paginação",
                "description": "Informações sobre a paginação dos resultados",
                "properties": {
                    "current_page": {
                        "description": "Página atual",
                        "type": "integer"
                    },
                    "per_page": {
                        "description": "Itens por página",
                        "type": "integer"
                    },
                    "total": {
                        "description": "Total de itens",
                        "type": "integer"
                    },
                    "total_pages": {
                        "description": "Total de páginas",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}